header: title, cartIcon & restaurants avaible
