+++ src/App.js
@@ -1,7 +1,6 @@
-import './App.css';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Home from './components/Home';
-import Header from './components/Header';
+import Header from './components/header/Header';
 
 function App() {
   return (
