+++ src/index.css
@@ -1,8 +1,27 @@
+@font-face {
+  font-family: 'cera-pro';
+  src: url('./assets/fonts/cerapro-bold.otf');
+  font-weight: bold;
+  font-style: normal;
+}
+
+@font-face {
+  font-family: 'cera-pro';
+  src: url('./assets/fonts/cerapro-medium.otf');
+  font-weight: medium;
+  font-style: normal;
+}
+
+@font-face {
+  font-family: 'cera-pro';
+  src: url('./assets/fonts/cerapro-light.otf');
+  font-weight: light;
+  font-style: normal;
+}
+
 body {
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
+  font-family: "cera-pro", sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
